requirements.txt
#for i in $(cat requirements.txt); do pip install $i; done
#for i in $(cat requirements.txt); do pip install --upgrade $i; done
# For testing
pytest
pytest-dotenv

# third-party (libraries)
rake_nltk
ipython

# linting stubs
types-requests
types-setuptools

# linting
black
isort
mypy
flake8
flake8-docstrings
pylint

git+https://github.com/openai/whisper.git
pyparsing
newspaper3k
requests
selenium
beautifulsoup4
ffmpeg-python
PyPDF2
tiktoken
moviepy
gradio
IPython
pydub
torch
wget
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
tqdm
nltk
gensim
# spacy
transformers
torch
pytube
flask
llama-hub>=0.0.79.post1
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-agent-openai
llama-index-tools-bing-search
llama-index-tools-weather
llama-index-tools-google
llama-index-tools-wikipedia
pyowm
llama-index
openai
langchain>=0.0.329
youtube_transcript_api
supabase
docx2txt
pypdf
cohere
groq
google.generativeai
pinecone
lxml_html_clean
nvidia-riva-client
soundfile
sounddevice
gpt_researcher
